Move LLMResponse to types
