+++ packages/server/src/sdk/workspace/ai/llm/bbai.ts
@@ -9,8 +9,7 @@ import { quotas } from "@budibase/pro"
 import { wrapLanguageModel } from "ai"
 import { TransformStream } from "node:stream/web"
 import { context } from "@budibase/backend-core"
-import { ReasoningEffort } from "@budibase/types"
-import { LLMResponse } from "."
+import { LLMResponse, ReasoningEffort } from "@budibase/types"
 import environment from "../../../../environment"
 
 interface OpenAIUsage {
