+++ packages/server/src/sdk/workspace/ai/llm/index.ts
@@ -1,6 +1,5 @@
-import type { EmbeddingModelV3, LanguageModelV3 } from "@ai-sdk/provider"
 import { HTTPError, env } from "@budibase/backend-core"
-import { BUDIBASE_AI_PROVIDER_ID } from "@budibase/types"
+import { BUDIBASE_AI_PROVIDER_ID, LLMResponse } from "@budibase/types"
 import tracer from "dd-trace"
 import sdk from "../../.."
 import { createBBAIClient } from "./bbai"
@@ -8,18 +7,6 @@ import { createLiteLLMOpenAI } from "./litellm"
 
 export * as bbai from "./bbai"
 export * from "./utils"
-
-export interface LLMResponse {
-  chat: LanguageModelV3
-  embedding: EmbeddingModelV3
-  providerOptions?: (hasTools: boolean) =>
-    | {
-        openai: {
-          parallelToolCalls: boolean
-        }
-      }
-    | undefined
-}
 
 export async function createLLM(
   configId: string,
